requirements.txt
fastapi
uvicorn
pandas
pydantic
langchain
groq
serpapi
sqlalchemy
requests
matplotlib
seaborn
jinja2
weasyprint
chromadb
huggingface-hub
sentence-transformers
pytest
python-dotenv
langchain-core
langchain-text-splitters
langsmith
pypdf